+++ lib/services/permission_service.dart
@@ -1,6 +1,7 @@
 import 'package:permission_handler/permission_handler.dart';
 
-abstract class SPWPermissionService {
+abstract class SPWPermissionService<T extends SPWPermissionService<T>> {
   Future<PermissionStatus> checkStatus();
   Future<PermissionStatus> request();
+  T get instance => this as T;
 }